migrate consensus to collections
