+++ baseapp/params_legacy.go
@@ -140,7 +140,7 @@ func GetConsensusParams(ctx sdk.Context, paramStore LegacyParamStore) *cmtproto.
 
 func MigrateParams(ctx sdk.Context, lps LegacyParamStore, ps ParamStore) {
 	if cp := GetConsensusParams(ctx, lps); cp != nil {
-		ps.Set(ctx, cp)
+		ps.Set(ctx, *cp)
 	} else {
 		ctx.Logger().Info("warning: consensus parameters are undefined; skipping migration")
 	}
