--- x/consensus/keeper/grpc_query.go
@@ -1,35 +1,0 @@
-package keeper
-
-import (
-	"context"
-
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/consensus/types"
-)
-
-var _ types.QueryServer = Querier{}
-
-// Querier is used as Keeper will have duplicate methods if used directly, and gRPC names take precedence over keeper
-type Querier struct {
-	Keeper
-}
-
-// NewQuerier constructor for the Querier struct
-func NewQuerier(keeper Keeper) Querier {
-	return Querier{Keeper: keeper}
-}
-
-// Params queries params of consensus module
-func (k Querier) Params(ctx context.Context, _ *types.QueryParamsRequest) (*types.QueryParamsResponse, error) {
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
-
-	params, err := k.Keeper.Get(sdkCtx)
-	if err != nil {
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-
-	return &types.QueryParamsResponse{Params: params}, nil
-}
