+++ x/consensus/keeper/keeper.go
@@ -3,29 +3,40 @@ package keeper
 import (
 	"context"
 
-	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
-
+	"cosmossdk.io/collections"
 	storetypes "cosmossdk.io/core/store"
+	"cosmossdk.io/errors"
+	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
+	cmttypes "github.com/cometbft/cometbft/types"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/x/consensus/exported"
 	"github.com/cosmos/cosmos-sdk/x/consensus/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
-var _ exported.ConsensusParamSetter = (*Keeper)(nil)
+var (
+	ParamsPrefix = collections.NewPrefix(0)
+	StoreKey     = "consensus"
+)
 
 type Keeper struct {
 	storeService storetypes.KVStoreService
 	cdc          codec.BinaryCodec
 
 	authority string
+	Params    collections.Item[cmtproto.ConsensusParams]
 }
 
 func NewKeeper(cdc codec.BinaryCodec, storeService storetypes.KVStoreService, authority string) Keeper {
+	sb := collections.NewSchemaBuilder(storeService)
 	return Keeper{
 		storeService: storeService,
 		cdc:          cdc,
 		authority:    authority,
+
+		Params: collections.NewItem(sb, ParamsPrefix, StoreKey, codec.CollValue[cmtproto.ConsensusParams](cdc)),
 	}
 }
 
@@ -33,30 +44,55 @@ func (k *Keeper) GetAuthority() string {
 	return k.authority
 }
 
-// Get gets the consensus parameters
-func (k *Keeper) Get(ctx context.Context) (*cmtproto.ConsensusParams, error) {
-	store := k.storeService.OpenKVStore(ctx)
+// Querier
 
-	bz, err := store.Get(types.ParamStoreKeyConsensusParams)
+var _ types.QueryServer = Querier{}
+
+// Querier is used as Keeper will have duplicate methods if used directly, and gRPC names take precedence over keeper
+type Querier struct {
+	Keeper
+}
+
+// NewQuerier constructor for the Querier struct
+func NewQuerier(keeper Keeper) Querier {
+	return Querier{Keeper: keeper}
+}
+
+// Params queries params of consensus module
+func (k Querier) Params(ctx context.Context, _ *types.QueryParamsRequest) (*types.QueryParamsResponse, error) {
+	params, err := k.Keeper.Params.Get(ctx)
 	if err != nil {
+		return nil, status.Error(codes.Internal, err.Error())
+	}
+
+	return &types.QueryParamsResponse{Params: &params}, nil
+}
+
+// MsgServer
+
+type msgServer struct {
+	Keeper
+}
+
+var _ types.MsgServer = msgServer{}
+
+// NewMsgServerImpl returns an implementation of the bank MsgServer interface
+// for the provided Keeper.
+func NewMsgServerImpl(keeper Keeper) types.MsgServer {
+	return &msgServer{Keeper: keeper}
+}
+
+func (k msgServer) UpdateParams(ctx context.Context, req *types.MsgUpdateParams) (*types.MsgUpdateParamsResponse, error) {
+	if k.GetAuthority() != req.Authority {
+		return nil, errors.Wrapf(govtypes.ErrInvalidSigner, "invalid authority; expected %s, got %s", k.GetAuthority(), req.Authority)
+	}
+
+	consensusParams := req.ToProtoConsensusParams()
+	if err := cmttypes.ConsensusParamsFromProto(consensusParams).ValidateBasic(); err != nil {
 		return nil, err
 	}
 
-	cp := &cmtproto.ConsensusParams{}
-	if err := k.cdc.Unmarshal(bz, cp); err != nil {
-		return nil, err
-	}
+	k.Params.Set(ctx, consensusParams)
 
-	return cp, nil
-}
-
-func (k *Keeper) Has(ctx context.Context) (bool, error) {
-	store := k.storeService.OpenKVStore(ctx)
-	return store.Has(types.ParamStoreKeyConsensusParams)
-}
-
-// Set sets the consensus parameters
-func (k *Keeper) Set(ctx context.Context, cp *cmtproto.ConsensusParams) error {
-	store := k.storeService.OpenKVStore(ctx)
-	return store.Set(types.ParamStoreKeyConsensusParams, k.cdc.MustMarshal(cp))
+	return &types.MsgUpdateParamsResponse{}, nil
 }
