+++ x/consensus/keeper/keeper_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
+	cmttypes "github.com/cometbft/cometbft/types"
 	"github.com/stretchr/testify/suite"
 
 	storetypes "cosmossdk.io/store/types"
@@ -15,7 +16,7 @@ import (
 	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	consensusparamkeeper "github.com/cosmos/cosmos-sdk/x/consensus/keeper"
-	consensusparamtypes "github.com/cosmos/cosmos-sdk/x/consensus/types"
+	"github.com/cosmos/cosmos-sdk/x/consensus/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
@@ -24,12 +25,12 @@ type KeeperTestSuite struct {
 	ctx                   sdk.Context
 	consensusParamsKeeper *consensusparamkeeper.Keeper
 
-	queryClient consensusparamtypes.QueryClient
-	msgServer   consensusparamtypes.MsgServer
+	queryClient types.QueryClient
+	msgServer   types.MsgServer
 }
 
 func (s *KeeperTestSuite) SetupTest() {
-	key := storetypes.NewKVStoreKey(consensusparamtypes.StoreKey)
+	key := storetypes.NewKVStoreKey(consensusparamkeeper.StoreKey)
 	testCtx := testutil.DefaultContextWithDB(s.T(), key, storetypes.NewTransientStoreKey("transient_test"))
 	ctx := testCtx.Ctx.WithBlockHeader(cmtproto.Header{})
 	encCfg := moduletestutil.MakeTestEncodingConfig()
@@ -40,13 +41,124 @@ func (s *KeeperTestSuite) SetupTest() {
 	s.ctx = ctx
 	s.consensusParamsKeeper = &keeper
 
-	consensusparamtypes.RegisterInterfaces(encCfg.InterfaceRegistry)
+	types.RegisterInterfaces(encCfg.InterfaceRegistry)
 	queryHelper := baseapp.NewQueryServerTestHelper(ctx, encCfg.InterfaceRegistry)
-	consensusparamtypes.RegisterQueryServer(queryHelper, consensusparamkeeper.NewQuerier(keeper))
-	s.queryClient = consensusparamtypes.NewQueryClient(queryHelper)
+	types.RegisterQueryServer(queryHelper, consensusparamkeeper.NewQuerier(keeper))
+	s.queryClient = types.NewQueryClient(queryHelper)
 	s.msgServer = consensusparamkeeper.NewMsgServerImpl(keeper)
 }
 
 func TestKeeperTestSuite(t *testing.T) {
 	suite.Run(t, new(KeeperTestSuite))
+}
+
+func (s *KeeperTestSuite) TestGRPCQueryConsensusParams() {
+	defaultConsensusParams := cmttypes.DefaultConsensusParams().ToProto()
+
+	testCases := []struct {
+		msg      string
+		req      types.QueryParamsRequest
+		malleate func()
+		response types.QueryParamsResponse
+		expPass  bool
+	}{
+		{
+			"success",
+			types.QueryParamsRequest{},
+			func() {
+				input := &types.MsgUpdateParams{
+					Authority: s.consensusParamsKeeper.GetAuthority(),
+					Block:     defaultConsensusParams.Block,
+					Validator: defaultConsensusParams.Validator,
+					Evidence:  defaultConsensusParams.Evidence,
+				}
+				s.msgServer.UpdateParams(s.ctx, input)
+			},
+			types.QueryParamsResponse{
+				Params: &cmtproto.ConsensusParams{
+					Block:     defaultConsensusParams.Block,
+					Validator: defaultConsensusParams.Validator,
+					Evidence:  defaultConsensusParams.Evidence,
+					Version:   defaultConsensusParams.Version,
+				},
+			},
+			true,
+		},
+	}
+
+	for _, tc := range testCases {
+		s.Run(tc.msg, func() {
+			s.SetupTest() // reset
+
+			tc.malleate()
+			res, err := s.queryClient.Params(s.ctx, &tc.req)
+
+			if tc.expPass {
+				s.Require().NoError(err)
+				s.Require().NotNil(res)
+				s.Require().Equal(tc.response.Params, res.Params)
+			} else {
+				s.Require().Error(err)
+				s.Require().Nil(res)
+			}
+		})
+	}
+}
+
+func (s *KeeperTestSuite) TestUpdateParams() {
+	defaultConsensusParams := cmttypes.DefaultConsensusParams().ToProto()
+	testCases := []struct {
+		name      string
+		input     *types.MsgUpdateParams
+		expErr    bool
+		expErrMsg string
+	}{
+		{
+			name: "valid params",
+			input: &types.MsgUpdateParams{
+				Authority: s.consensusParamsKeeper.GetAuthority(),
+				Block:     defaultConsensusParams.Block,
+				Validator: defaultConsensusParams.Validator,
+				Evidence:  defaultConsensusParams.Evidence,
+			},
+			expErr:    false,
+			expErrMsg: "",
+		},
+		{
+			name: "invalid  params",
+			input: &types.MsgUpdateParams{
+				Authority: s.consensusParamsKeeper.GetAuthority(),
+				Block:     &cmtproto.BlockParams{MaxGas: -10, MaxBytes: -10},
+				Validator: defaultConsensusParams.Validator,
+				Evidence:  defaultConsensusParams.Evidence,
+			},
+			expErr:    true,
+			expErrMsg: "block.MaxBytes must be greater than 0. Got -10",
+		},
+		{
+			name: "invalid authority",
+			input: &types.MsgUpdateParams{
+				Authority: "invalid",
+				Block:     defaultConsensusParams.Block,
+				Validator: defaultConsensusParams.Validator,
+				Evidence:  defaultConsensusParams.Evidence,
+			},
+			expErr:    true,
+			expErrMsg: "invalid authority",
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		s.Run(tc.name, func() {
+			s.SetupTest()
+			_, err := s.msgServer.UpdateParams(s.ctx, tc.input)
+			if tc.expErr {
+				s.Require().Error(err)
+				s.Require().Contains(err.Error(), tc.expErrMsg)
+			} else {
+				s.Require().NoError(err)
+			}
+		})
+	}
 }
