+++ x/consensus/module.go
@@ -168,7 +168,7 @@ func ProvideModule(in ConsensusInputs) ConsensusOutputs {
 	k := keeper.NewKeeper(in.Cdc, in.StoreService, authority.String())
 	m := NewAppModule(in.Cdc, k)
 	baseappOpt := func(app *baseapp.BaseApp) {
-		app.SetParamStore(&k)
+		app.SetParamStore(k.Params)
 	}
 
 	return ConsensusOutputs{
