+++ x/consensus/types/keys.go
@@ -7,5 +7,3 @@ const (
 	// StoreKey defines the module's store key.
 	StoreKey = ModuleName
 )
-
-var ParamStoreKeyConsensusParams = []byte("Consensus")
